Don't page through non posts
I wished to add a privacy policy but it was showing up in the list of
articles on the main page. I'm fixing this problem in the theme by
restricting it to only page through anything in the section "posts".

This syntax is very unusual to me so figuring out the right combination
was tricky. The docs also reference `.Paginate` and use that to create
a variable called `$paginator`, but this theme uses `.Paginator`.
I didn't intially spot that one character diference, but it seperates
working code from non working code. :D